+++ tests/tests.rs
@@ -18,3 +18,4 @@ mod test_builtins;
 mod test_ints;
 mod test_typedef;
 mod test_keywords;
+mod test_prefix;