+++ Changelog.md
@@ -15,6 +15,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
   instead of `std::os::raw`. See `--ctypes-prefix` (#316)
 - Add the possibility to use `core` instead of `std` in the generated source.
   See `--use-core` (#316)
+- Possibility to remove a prefix from all symbols, like `libfoo_`. See
+  `--remove-prefix` (#93)
 
 ### Fixed
 - Now correctly handle variable names like "u8".